Portable header-only file "magicavoxel_loader.h" for MagicaVoxel loader example.
