+++ projects/VS2019/examples/models_mesh_magicavoxel_loading.vcxproj
@@ -381,6 +381,9 @@
   <ItemGroup>
     <ClCompile Include="..\..\..\examples\models\models_magicavoxel_loading.c" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\..\..\examples\models\magicavoxel_loader.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
